moved whale and playground to /api
