+++ hooks/api/usePlaygroundRpcClient.ts
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux'
 import { PlaygroundApiClient, PlaygroundRpcClient } from '@defichain/playground-api-client'
-import { RootState } from '../store'
-import { PlaygroundApiState } from '../store/network'
+import { RootState } from '../../store'
+import { PlaygroundApiState } from '../../store/network'
 
 export function usePlaygroundRpcClient (): PlaygroundRpcClient {
   const playground = useSelector<RootState, PlaygroundApiState | undefined>(state => state.network.playground)
